requirements.txt
numpy
pygame
numpy-quaternion
